added new TestSetup() and NewConfig() funcs; experimental APIs
These may or may not be kept, but we should use something like
these or find better alternatives.
