+++ cmd/protoc-gen-gorums/dev/config.go
@@ -1,6 +1,8 @@
 package dev
 
-import "fmt"
+import (
+	"fmt"
+)
 
 // A Configuration represents a static set of nodes on which quorum remote
 // procedure calls may be invoked.
@@ -11,6 +13,21 @@ type Configuration struct {
 	mgr   *Manager
 	qspec QuorumSpec
 	errs  chan GRPCError
+}
+
+// NewConfig returns a configuration for the given node addresses and quorum spec.
+// The returned func() must be called to close the underlying connections.
+// This is experimental API.
+func NewConfig(addrs []string, qspec QuorumSpec, opts ...ManagerOption) (*Configuration, func(), error) {
+	man, err := NewManager(addrs, opts...)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to create manager: %v", err)
+	}
+	c, err := man.NewConfiguration(man.NodeIDs(), qspec)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to create configuration: %v", err)
+	}
+	return c, func() { man.Close() }, nil
 }
 
 // ID reports the identifier for the configuration.
